Updated the Cassandra driver to 3.0.1 and Cassandra Unit to 3.0.0.1
Modified project dependencies to point to the Java Cassandra driver version 3.0.1. Updated the dependencies
well to point to Cassandra Unit 3.0.0.1. Updated the CassandraEventLog class to conform to the new driver
by removing the QueryTimeoutException and replacing it with OperationTimeoutException and
WriteTimeoutException.

This will allow for users of Eventuate to leverage Cassandra 3.x since it requires the Java 3.0+ driver.