Support binary payload serializer for DurableEvents
Next to the default DelegatingDurableEventSerializer there is an additional
DurableEventSerializerWithBinaryPayload that does not use akka's
serialization mechanism for de-/serializing payloads but
de-/serializes payloads from/into BinaryPayload instances. This can
be useful in scenarios where an Eventuate application is just used for
event forwarding (through replication) and only handles the
payload of events based on their metadata (like the manifest).
With BinaryPayloadManifestFilter there is also a ReplicationFilter
that operates on those BinaryPayloads and passes only those
whose manifest matches a given regexp.