Upgrade to Akka 2.5.7
Akka 2.5.7 introduced the following incompatible changes in the Java-API
- ReceiveBuilder:
  - explicit create step required
  - builds `AbstractActor.Receive` instead of `Actor.Receive`
- Match.match: Changes in type parameters

To adjust to these changes the Java-API of Eventuate is streamlined
and mirrors the new Akka Java-API by removing all `setOn...`-methods
used to defined the actor's behavior and replaces these methods with
`createOn...`-variants which can be used to define custom behavior
by returning the behavior definition wrapped in an instance of
`AbstractActor.Receive`.

New Java-specific accessor methods are introduced to access the
properties of any event-sourced component with the Java-API.

`AbstractEventsourcedProcessor` supports behavior definition by overriding
the `createOnProcess` method which facilitates a Java-specific `Process` type
to define custom behavior.
The `ProcessBuilder` may be used to create instances of type `Process`.

A Java-specific `BehaviorContext` is added to the Java-API of al all
event-sourced components which can be retrieved by calling `getContext()`.