Update Cassandra index on AdjustEventLogClock
When writing the adjusted event log clock after disaster recovery also
update the index up to the last sequence nr to ensure that the written
event log clock snapshot is consistent with the index.

Before the index-actor informed the log-actor with a message about
an up-to-date index. This has been replaces by a promise so that
its future can be returned. The alternative of using and ask would
have required to specify an timeout, however it is hard to define
a reasonable default for this timeout.

Closes #393