Fix failing LogEventDispatcherSpec
`LogEventDispatcherSpec` used a static delay between the start of a `LogEventDispatcher`
and delivery of the test messages. This delay could lead to a situation where messages
where sent to the `EventLogDispatcher` but the initialization process was not yet completed
resulting in a loss of the sent messages.

Implement a delayed retry-mechanism for sending events to the `LogEventDispatcher`
if event-delivery fails with a 'ReplyException.NoHandlers' indicating that
initialization of the `LogEventDispatcher` has not yet completed.

Closes #377